fix: resolve all build warnings and errors
- remove imgui backends requiring external libraries (allegro, sdl, vulkan, etc)
- fix wchar to string conversion in http.hpp
- replace luaL_loadbuffer with luau_compile/load in filesystem.hpp
- fix api_incr_top macro issues in environment.hpp
- fix nullptr to ImTextureID cast and std::max macro conflict in drawing.hpp
- fix setupvalue macro conflict and add missing includes in debug.hpp
- fix UDATA_META_DEC and luaC_barriert issues in metatable.hpp
- fix lua_newuserdatadtor and closure member access in closures.hpp
- update all lua_pushcclosure to lua_pushcclosurek across codebase
- fix c4312 warnings in cpr/session.cpp
- fix size_t to unsigned int warnings in constantfolding.cpp

+++ YuB-X-Module/Dependencies/cpr/src/session.cpp
@@ -195,7 +195,7 @@ void Session::Impl::SetMultipart(Multipart&& multipart) {
                 formdata.push_back(
                         {CURLFORM_COPYCONTENTS, reinterpret_cast<const char*>(part.data)});
                 formdata.push_back(
-                        {CURLFORM_CONTENTSLENGTH, reinterpret_cast<const char*>(part.datalen)});
+                        {CURLFORM_CONTENTSLENGTH, reinterpret_cast<const char*>(static_cast<uintptr_t>(part.datalen))});
             } else if (part.is_file) {
                 formdata.push_back({CURLFORM_FILE, part.value.data()});
             } else {
@@ -227,7 +227,7 @@ void Session::Impl::SetMultipart(const Multipart& multipart) {
                 formdata.push_back({CURLFORM_BUFFER, part.value.data()});
                 formdata.push_back({CURLFORM_BUFFERPTR, reinterpret_cast<const char*>(part.data)});
                 formdata.push_back(
-                        {CURLFORM_BUFFERLENGTH, reinterpret_cast<const char*>(part.datalen)});
+                        {CURLFORM_BUFFERLENGTH, reinterpret_cast<const char*>(static_cast<uintptr_t>(part.datalen))});
             } else if (part.is_file) {
                 formdata.push_back({CURLFORM_FILE, part.value.data()});
             } else {
